requirements.txt
streamlit
opencv-python-headless
numpy
pillow
matplotlib
roboflow
ultralytics